+++ packages/build-user-config/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.3.6
 
 - [fix] SPA `route.targets` field
+- [fix] regexp for inject core-js
+- [fix] require order of polyfill
 
 ## 0.3.5