+++ packages/build-user-config/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.3.6
+
+- [feat] use trusted-cert to generate ca
+- [feat] add `experiments` to userConfig
+- [fix] SPA `route.targets` field
+- [fix] regexp for inject core-js
+- [fix] require order of polyfill
+
 ## 0.3.5
 
 - [feat] add `props.pageConfig` in Rax App SPA page component